refactor xplorers-ai
