+++ package.json
@@ -33,6 +33,7 @@
     "marked": "^15.0.6",
     "next": "14.2.3",
     "next-cloudinary": "^6.16.0",
+    "notistack": "^3.0.1",
     "openai": "^4.78.1",
     "react": "^18",
     "react-dom": "^18",
@@ -68,5 +69,6 @@
     "*.{json,md,yml}": [
       "prettier --write"
     ]
-  }
+  },
+  "packageManager": "pnpm@9.2.0+sha512.98a80fd11c2e7096747762304106432b3ddc67dcf54b5a8c01c93f68a2cd5e05e6821849522a06fb76284d41a2660d5e334f2ee3bbf29183bf2e739b1dafa771"
 }
