+++ src/hooks/useXplorersAI.ts
@@ -1,18 +1,18 @@
-"use client";
-import React, { ChangeEvent, useRef, useState, MouseEvent } from "react";
 import { callAzureOpenAI } from "@/app/api/chat/ai";
-import { ChatMessage, FileUploadEvent } from "@/models/XploresAI";
-import { XplorersAIContext } from "@/context/XplorersAIContext";
+import { ChatMessage } from "@/models/XploresAI";
+import { useSnackbar } from "notistack";
+import { ChangeEvent, useRef, useState } from "react";
 
-const XplorersAIContextProvider: React.FC<{
-  children: React.ReactNode;
-}> = ({ children }) => {
+const useXplorersAI = () => {
+  const { enqueueSnackbar } = useSnackbar();
   const [value, setValue] = useState<string>("");
   const [loading, setLoading] = useState<boolean>(false);
   const textareaRef = useRef<HTMLTextAreaElement | null>(null);
+
   const [messages, setMessages] = useState<ChatMessage[]>([
     { role: "user", content: "" },
   ]);
+
   const handleInputChange = (e: ChangeEvent<HTMLTextAreaElement>) => {
     const textValue = e.target.value;
     setValue(textValue);
@@ -23,14 +23,16 @@ const XplorersAIContextProvider: React.FC<{
     }
   };
 
-  const handleFileUpload = (e: FileUploadEvent) => {
+  const handleFileUpload = (e: React.ChangeEvent<HTMLInputElement>) => {
     const file = e.target.files ? e.target.files[0] : null;
     if (file) {
       console.debug("File uploaded:", file);
     }
   };
-  const handleSubmit = async (e: MouseEvent<HTMLButtonElement>) => {
+
+  const handleSubmit = async (e: React.MouseEvent<HTMLButtonElement>) => {
     setLoading(true);
+
     if (value.trim()) {
       if (textareaRef.current) {
         textareaRef.current.style.height = "auto";
@@ -41,6 +43,7 @@ const XplorersAIContextProvider: React.FC<{
       role: "user",
       content: value,
     };
+
     setMessages((prev) => [...prev, userMessage]);
 
     try {
@@ -53,29 +56,25 @@ const XplorersAIContextProvider: React.FC<{
 
       setMessages((prev) => [...prev, systemMessage]);
       setValue("");
-
       setLoading(false);
     } catch (error) {
       console.debug("AI Call Failed", error);
+      setLoading(false);
+      enqueueSnackbar({
+        message: "Something went wrong! Please try again later.",
+        variant: "error",
+      });
     }
   };
 
-  const getMessages = () => messages;
-
-  return (
-    <XplorersAIContext.Provider
-      value={{
-        value,
-        loading,
-        handleInputChange,
-        handleFileUpload,
-        handleSubmit,
-        getMessages,
-      }}
-    >
-      {children}
-    </XplorersAIContext.Provider>
-  );
+  return {
+    value,
+    loading,
+    handleInputChange,
+    handleFileUpload,
+    handleSubmit,
+    messages,
+  };
 };
 
-export default XplorersAIContextProvider;
+export default useXplorersAI;
